fix potentially wrong request for size to malloc in zlib
item and size are multiplied, but are potentially signed. If an overflow turns into an underflow, this malloc/calloc are likely doomed.

This comes from #47. 

I don't like this patch and may not submit this at all depending on what upstream does. This is mostly to see how presubmit like it.

+++ zlib/zutil.c
@@ -310,8 +310,8 @@ voidpf ZLIB_INTERNAL zcalloc(opaque, items, size)
     unsigned size;
 {
     (void)opaque;
-    return sizeof(uInt) > 2 ? (voidpf)malloc(items * size) :
-                              (voidpf)calloc(items, size);
+    return sizeof(uInt) > 2 ? (voidpf)malloc((size_t)items * (size_t)size) :
+                              (voidpf)calloc((size_t)items, (size_t)size);
 }
 
 void ZLIB_INTERNAL zcfree(opaque, ptr)
